cloud-functions/bq_to_firestore/requirements.txt
google-cloud-core
google-cloud-bigquery
firebase_admin